pyproject.toml
[tool.poetry]
name = "llm-trajectory-dynamics"
version = "2.0.0"
description = "LLM trajectory dynamics analysis framework"
authors = ["xiangli"]
readme = "README.md"
packages = [{ include = "lmd" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"

# Hugging Face
transformers = "4.45.*"
accelerate = "0.34.*"
datasets = "2.21.*"
tokenizers = "0.20.*"
huggingface-hub = "0.24.*"
safetensors = "0.4.*"

# Data Storation
zarr = "2.18.*"
numcodecs = "0.13.*"
pandas = "2.2.*"
pyarrow = "17.*"
numpy = "1.26.*"
jsonlines = "4.0.*"

# computing
scipy = "1.14.*"
scikit-learn = "1.5.*"
numba = "0.60.*"
joblib = "1.4.*"

# viz
matplotlib = "3.9.*"
seaborn = "0.13.*"

# Tools
tqdm = "4.66.*"
pyyaml = "6.0.*"
hydra-core = "1.3.*"
omegaconf = "2.3.*"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.*"
pytest-cov = "5.0.*"
black = "24.8.*"
isort = "5.13.*"
ipykernel = "6.29.*"
jupyter = "1.0.*"
notebook = "7.2.*"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"